+++ rumbas/src/data/default.rs
@@ -103,6 +103,9 @@ impl DefaultFileType {
                 Some("questionpart.gapfill.gap.pattern_match") => Some(
                     DefaultFileType::QuestionPartGapFillGap(QuestionPartType::PatternMatch),
                 ),
+                Some("questionpart.gapfill.gap.choose_one") => Some(
+                    DefaultFileType::QuestionPartGapFillGap(QuestionPartType::ChooseOne),
+                ),
                 _ => None,
             },
             None => None,
@@ -482,6 +485,13 @@ pub fn combine_with_default_files(path: &Path, exam: &mut Exam) {
                                                                             if let (
                                                                     QuestionPart::PatternMatch(_),
                                                                     QuestionPart::PatternMatch(_),
+                                                                ) = (&p, &gap)
+                                                                {
+                                                                    gap.overwrite(&p.clone())
+                                                                }
+                                                                            if let (
+                                                                    QuestionPart::ChooseOne(_),
+                                                                    QuestionPart::ChooseOne(_),
                                                                 ) = (&p, &gap)
                                                                 {
                                                                     gap.overwrite(&p.clone())